save js in seperate file
