+++ .pre-commit-config.yaml
@@ -46,16 +46,6 @@ repos:
     hooks:
       - id: oxipng
         stages: [manual]
-  - repo: https://github.com/pre-commit/mirrors-eslint
-    rev: v9.13.0
-    hooks:
-      - id: eslint
-        args: ['-c', 'panel/.eslintrc.js', 'panel/*.ts', 'panel/models/**/*.ts', '--fix']
-        additional_dependencies:
-          - 'eslint@8.57.0'
-          - '@stylistic/eslint-plugin@1.6.3'
-          - '@typescript-eslint/eslint-plugin@7.2.0'
-          - '@typescript-eslint/parser@7.2.0'
   - repo: https://github.com/pre-commit/mirrors-mypy
     rev: v1.13.0
     hooks:
